Update
- use `isnan()` instead of `isnanf()`
- Add litmus tests for add identity
- Use `float` (f32) instead of `double` (f64) in `Float`
- Fix logics related to NaN in `Float`
Add litmus tests for fadd NaN, x
- Revert "changing double to float"
- Encode Inf arithmetic properties in `aop::fpAdd`
- Remove redundant parantheses
- Do not encode signs of NaN

+++ src/abstractops.h
@@ -24,9 +24,9 @@ AbsLevelDot getAbstractionLevelOfDot();
 bool getAddAssociativity();
 
 smt::Sort fpSort();
-smt::Expr fpConst(float f);
+smt::Expr fpConst(double f);
 // Return the set of possible FP constants for 'e'.
-std::vector<float> fpPossibleConsts(const smt::Expr &e);
+std::vector<double> fpPossibleConsts(const smt::Expr &e);
 
 smt::Expr mkZeroElemFromArr(const smt::Expr &arr);
 smt::Expr fpAdd(const smt::Expr &f1, const smt::Expr &f2);
